apps/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Application to consume bigquery data"
authors = ["Andres Garcia <jose.andres.gm29@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
google-cloud-bigquery = "^3.12.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
